requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
distro==1.9.0
dropbox==12.0.2
et_xmlfile==2.0.0
exceptiongroup==1.3.1
geographiclib==2.1
geopy==2.4.1
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
jiter==0.13.0
lxml==6.0.2
msal==1.34.0
numpy==2.4.2
oauthlib==3.3.1
openai==2.21.0
openpyxl==3.1.5
packaging==26.0
pandas==3.0.1
pillow==12.1.1
ply==3.11
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
PyMuPDF==1.27.1
pyparsing==3.3.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stone==3.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
